+++ api/v1alpha1/profilecatalogsource_types.go
@@ -28,7 +28,7 @@ type ProfileCatalogSourceSpec struct {
 	Profiles []ProfileCatalogEntry `json:"profiles,omitempty"`
 }
 
-// ProfileDescription defines details about a given profile.
+// ProfileCatalogEntry defines details about a given profile.
 type ProfileCatalogEntry struct {
 	// Tag
 	// +optional
@@ -37,7 +37,7 @@ type ProfileCatalogEntry struct {
 	Tag string `json:"tag,omitempty"`
 	// CatalogSource is the name of the catalog the profile is listed in
 	// +optional
-	CatalogSource string `json:"catalog,omitempty"`
+	CatalogSource string `json:"catalogSource,omitempty"`
 	// URL is the full URL path to the profile.yaml
 	// +optional
 	URL                string `json:"url,omitempty"`
@@ -51,7 +51,7 @@ type ProfileCatalogSourceStatus struct {
 // +kubebuilder:object:root=true
 // +kubebuilder:subresource:status
 
-// ProfileCatalogSource is the Schema for the profilecatalogsources API
+// ProfileCatalogSource is the Schema for the ProfileCatalogSources API
 type ProfileCatalogSource struct {
 	metav1.TypeMeta   `json:",inline"`
 	metav1.ObjectMeta `json:"metadata,omitempty"`
