+++ config/manager/kustomization.yaml
@@ -9,9 +9,3 @@ configMapGenerator:
 - files:
   - controller_manager_config.yaml
   name: manager-config
-apiVersion: kustomize.config.k8s.io/v1beta1
-kind: Kustomization
-images:
-- name: weaveworks/profiles-controller
-  newName: localhost:5000/profiles-controller
-  newTag: latest
